weather_bot/requirements.txt
pyTelegramBotAPI==4.22.1
requests==2.25.1
python-dotenv==1.0.1
